Separate build from runtime file inclusion and exclusion (Fixes #25) (#31)
New CLI args:

* `--include` run the linter on these path ignoring the includes and
excludes defined in the `build.zig` forcing these paths to be resolved
and linted (if they exist).
* `--exclude` exclude these paths from linting. This argument will be
used in conjunction with the excludes defined in the `build.zig` unless
used with `--include`.
* `--filter` used to filter the run to a specific set of already
resolved paths. Unlike `--include` this leaves the includes and excludes
defined in the `build.zig` as is.