Use aliases for `std.zig.Ast` and `zlinter.shims` to reduce line noise (#102)
Originally I was being verbose while things found their places. Now that
pieces are falling together I think it's time to be a little less
verbose. `shims` and `Ast` seem like a good case as they appear in a lot
of places.

So after a few unix commands and a few `git restore .` calls later we
have this. Maybe in time I'll extend to more but I think this is
probably enough.