Fix: Coverage timing out on master (#125)
Seems that it's getting stuck on dependencies, so temporarily just going
to manually fetch all dependencies (zls) and transitive dependencies.
It's a bit of a hack, if we need to do this long term, maybe worth
trying to upstream a fix to `zig build --fetch` to also fetch transitive
dependencies.